requirements.txt
Flask
transformers
torch
numpy
pandas
requests
pillow
pytesseract
